RavenDB-18778 Moving tests having long duration to SlowTests
